src/requirements.txt
anthropic>=0.26.0
pyyaml>=6.0
python-dotenv>=1.0
discord.py>=2.3.0
